backend/requirements.txt

fastapi
uvicorn
python-dotenv
requests
pandas
numpy
scikit-learn
inference
python-multipart
huggingface_hub